feat: add bulk parameters to scim configuration #1819

+++ jans-scim/model/src/main/java/io/jans/scim/model/scim2/provider/config/BulkConfig.java
@@ -27,7 +27,7 @@ public class BulkConfig {
             isRequired = true,
             mutability = AttributeDefinition.Mutability.READ_ONLY,
             type = AttributeDefinition.Type.INTEGER)
-	private long maxOperations;
+	private int maxOperations;
 
     @Attribute(description = " An integer value specifying the maximum payload size in bytes",
             isRequired = true,
@@ -46,7 +46,7 @@ public class BulkConfig {
 	 * @param maxOperations Specifies the maximum number of operations supported per bulk.
 	 * @param maxPayloadSize Specifies the maximum payload size in bytes supported per bulk.
 	 */
-	public BulkConfig(boolean supported, long maxOperations, long maxPayloadSize) {
+	public BulkConfig(boolean supported, int maxOperations, long maxPayloadSize) {
 		this.supported = supported;
 		this.maxOperations = maxOperations;
 		this.maxPayloadSize = maxPayloadSize;
@@ -64,7 +64,7 @@ public class BulkConfig {
 	 * Retrieves the maximum number of operations supported in a bulk.
 	 * @return The maximum number of operations.
 	 */
-	public long getMaxOperations() {
+	public int getMaxOperations() {
 		return maxOperations;
 	}
 
@@ -80,7 +80,7 @@ public class BulkConfig {
         this.supported = supported;
     }
 
-    public void setMaxOperations(long maxOperations) {
+    public void setMaxOperations(int maxOperations) {
         this.maxOperations = maxOperations;
     }
